\name{GDP.Per.head.dist.1995}
\alias{GDP.Per.head.dist.1995}


\title{annual Gross Domestic Product (GDP) per head across 15 European Union (EU) countries}

\description{Contains values of  the GDP/head distribution of 216 EU regions (the so called NUTS-2 level of the Eurostat categorization of territories within the EU for the year 1995.}

\usage{
  GDP.Per.head.dist.1995
}

\format{
A vector with 184 values of  the GDP/head distribution for 1995.
}

\source{
Monfort, P. (2008). Convergence of EU regions measures and evolution. EU short papers on regional research and indicators, Directorate-General for Regional Policy 1/2008.
}





\references{
  \href{https://ec.europa.eu/regional_policy/sources/docgener/work/200801_convergence.pdf}{Monfort, P. (2008). Convergence of EU regions measures and evolution. EU short papers on regional research and indicators}
    }




\seealso{
  \code{\link{GDP.Per.head.dist.2005}
  }
}

